Extract prompt building logic into separate file
